sample/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.1
django-filter==23.2
Pillow==9.5.0
sqlparse==0.4.4
typing-extensions==4.6.2
